# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PTIT eCommerce API",
        "description": "Tài liệu API cho PTIT eCommerce, bao gồm các luồng của Admin, Nhân viên và Khách hàng.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Current host"
        }
    ],
    "paths": {
...
        "/api/auth/profile": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Lấy thông tin profile người dùng",
                "operationId": "b70772ad1bb405cd5657bfefc25d31b2",
                "responses": {
                    "200": {
                        "description": "Thông tin người dùng",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/profile": {
            "get": {
                "tags": [
                    "User - Profile"
                ],
                "summary": "Xem thông tin tài khoản (web)",
                "description": "Yêu cầu phiên đăng nhập web. Endpoint chủ yếu để mô tả luồng người dùng trên Swagger.",
                "operationId": "90bc1ecd527a0d8c54f074880c21ccbb",
                "responses": {
                    "200": {
                        "description": "Trang profile người dùng"
                    }
                }
            },
            "put": {
                "tags": [
                    "User - Profile"
                ],
                "summary": "Cập nhật thông tin người dùng",
                "description": "Cập nhật tên, email, số điện thoại, địa chỉ, mật khẩu và avatar cho người dùng hiện tại.",
                "operationId": "2f80a564eef93f304c6dc0377da44e46",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Nguyễn Văn A"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    },
                                    "phone_number": {
                                        "type": "string",
                                        "example": "0123456789"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Hà Nội"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "new-password"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "new-password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "Redirect sau khi cập nhật thành công"
                    },
                    "422": {
                        "description": "Dữ liệu không hợp lệ"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}